+++ apps/cold-api/src/platform/modules/resources/organizations/organization_compliance/organization_compliance_responses/organization_compliance_ai_response_files/organization_compliance_ai_response_files.service.ts
@@ -1,28 +1,29 @@
 import { Injectable } from '@nestjs/common';
 import { BaseWorker, ComplianceAiResponseFilesRepository, IAuthenticatedUser } from '@coldpbc/nest';
+import { organizations } from '@prisma/client';
 
 @Injectable()
 export class OrganizationComplianceAiResponseFilesService extends BaseWorker {
   constructor(readonly repository: ComplianceAiResponseFilesRepository) {
     super(OrganizationComplianceAiResponseFilesService.name);
   }
-  create(orgId: string, complianceName: string, aiResponseId: string, aiResponseFileData: any, user: IAuthenticatedUser) {
-    return this.repository.createAiResponseFile(orgId, complianceName, aiResponseId, aiResponseFileData, user);
+  create(org: organizations, complianceName: string, aiResponseId: string, aiResponseFileData: any, user: IAuthenticatedUser) {
+    return this.repository.createAiResponseFile(org, complianceName, aiResponseId, aiResponseFileData, user);
   }
 
-  findAll(orgId: string, complianceName: string, aiResponseId: string, user: IAuthenticatedUser) {
-    return this.repository.findAllAiResponseFiles(orgId, complianceName, aiResponseId, user);
+  findAll(org: organizations, complianceName: string, aiResponseId: string, user: IAuthenticatedUser) {
+    return this.repository.findAllAiResponseFiles(org, complianceName, aiResponseId, user);
   }
 
-  findOne(orgId: string, complianceName: string, aiResponseId: string, id: string, user: IAuthenticatedUser) {
-    return this.repository.findOneAiResponseFile(orgId, complianceName, aiResponseId, id, user);
+  findOne(org: organizations, complianceName: string, aiResponseId: string, id: string, user: IAuthenticatedUser) {
+    return this.repository.findOneAiResponseFile(org, complianceName, aiResponseId, id, user);
   }
 
-  update(orgId: string, complianceName: string, aiResponseId: string, id: string, aiResponseFileData: any, user: IAuthenticatedUser) {
-    return this.repository.updateAiResponseFile(orgId, complianceName, aiResponseId, id, aiResponseFileData, user);
+  update(org: organizations, complianceName: string, aiResponseId: string, id: string, aiResponseFileData: any, user: IAuthenticatedUser) {
+    return this.repository.updateAiResponseFile(org, complianceName, aiResponseId, id, aiResponseFileData, user);
   }
 
-  remove(orgId: string, complianceName: string, aiResponseId: string, id: string, user: IAuthenticatedUser) {
-    return this.repository.removeAiResponseFile(orgId, complianceName, aiResponseId, id, user);
+  remove(org: organizations, complianceName: string, aiResponseId: string, id: string, user: IAuthenticatedUser) {
+    return this.repository.removeAiResponseFile(org, complianceName, aiResponseId, id, user);
   }
 }