+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+
 # [1.293.0](https://github.com/ColdPBC/cold-monorepo/compare/v1.292.0...v1.293.0) (2024-07-16)